using integers is more safe

+++ tests/cpp_caching/binary_operation.cc
@@ -1,3 +1,3 @@
-float op(float a, float b) {
+float op(int a, int b) {
     return a + b;
 }
