+++ tests/cpp_caching/binary_operation.h
@@ -1,1 +1,1 @@
-float op(float a, float b);
+float op(int a, int b);
