+++ tests/cpp_caching/op_user_lib.cc
@@ -1,5 +1,5 @@
 #include "binary_operation.h"
 
-float use_op() {
+int use_op() {
     return op(10, 2);
 }
